requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.37.31
botocore==1.37.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
